+++ apps/omsorgspengerutbetaling-soknad/src/app/i18n/appMessages.ts
@@ -61,6 +61,10 @@ const nb = {
     'resetMellomlagring.text.1': 'Dersom feilen vedvarer, kan du prøve å starte på nytt med et tom skjema.',
     'resetMellomlagring.startPåNytt': 'Start på nytt',
 
+    'page.ikkeTilgang.tekst':
+        'Du har ikke tilgang til denne siden. Dersom du er under 18 år, må en av foreldrene dine eller en foresatt skrive under på søknaden sammen med deg. Du må derfor fylle ut søknaden på papir og sende den i posten.',
+    'page.ikkeTilgang.lastNed': 'Her kan du laste ned papirsøknaden',
+
     'validation.barn.ingenBarn': 'Du må legge til minst ett barn for å kunne gå videre i søknaden.',
     'validation.harUtvidetRett.yesOrNoIsUnanswered':
         'Du må svare at du har fått ekstra omsorgsdager for barn fordi barnet har kronisk sykdom eller funksjonshemning eller ikke.',