+++ apps/omsorgspengerutbetaling-soknad/src/app/søknad/steps/dine-barn/DineBarnForm.tsx
@@ -1,13 +1,12 @@
 import { Heading } from '@navikt/ds-react';
 import React from 'react';
+import { RegistrertBarn, Søker } from '@navikt/sif-common';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 import FormBlock from '@navikt/sif-common-core-ds/src/atoms/form-block/FormBlock';
 import { FormikInputGroup, getTypedFormComponents, ValidationError } from '@navikt/sif-common-formik-ds';
 import getIntlFormErrorHandler from '@navikt/sif-common-formik-ds/src/validation/intlFormErrorHandler';
 import { AnnetBarn } from '@navikt/sif-common-forms-ds/src/forms/annet-barn/types';
 import { AppText, useAppIntl } from '../../../i18n';
-import { RegistrertBarn } from '../../../types/RegistrertBarn';
-import { Søker } from '../../../types/Søker';
 import { DineBarnFormFields, DineBarnFormValues } from './DineBarnStep';
 import AndreBarnPart from './parts/AndreBarnPart';
 import DineBarnStepIntro from './parts/DineBarnStepIntro';