+++ apps/omsorgspengerutbetaling-soknad/src/app/types/SøknadContextState.ts
@@ -1,7 +1,6 @@
+import { RegistrertBarn, Søker } from '@navikt/sif-common';
 import { DineBarnFormValues } from '../søknad/steps/dine-barn/DineBarnStep';
-import { RegistrertBarn } from './RegistrertBarn';
 import { StepId } from './StepId';
-import { Søker } from './Søker';
 import { SøknadRoutes } from './SøknadRoutes';
 import { Søknadsdata } from './søknadsdata/Søknadsdata';