+++ packages/sif-common/src/api/types/innsendingType.ts
@@ -1,0 +1,3 @@
+export enum InnsendingType {
+    'omsorgspenger_utbetaling_snf' = '/omsorgspenger-utbetaling-snf/innsending',
+}