+++ server/src/utils/reverseProxy.ts
@@ -2,6 +2,7 @@ import { getToken, requestTokenxOboToken } from '@navikt/oasis';
 import { Express, NextFunction, Request, Response } from 'express';
 import { createProxyMiddleware } from 'http-proxy-middleware';
 import config, { Service, verifyProxyConfigIsSet } from './serverConfig.js';
+import { v4 as uuid } from 'uuid';
 
 type ProxyOptions = {
     ingoingUrl: string;
@@ -39,6 +40,9 @@ export function addProxyHandler(server: Express, { ingoingUrl, outgoingUrl, scop
         async (request: Request, response: Response, next: NextFunction) => {
             if (process.env.NAIS_CLIENT_ID !== undefined) {
                 request.headers['X-K9-Brukerdialog'] = process.env.NAIS_CLIENT_ID;
+            }
+            if (request.headers['X-Correlation-ID'] === undefined) {
+                request.headers['X-Correlation-ID'] = uuid();
             }
             request.headers['X-Brukerdialog-Git-Sha'] = getCommitShaFromEnv();